+++ logic/src/test/java/org/itsallcode/whiterabbit/logic/service/plugin/PluginManagerTest.java
@@ -10,6 +10,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Optional;
 
 import org.itsallcode.whiterabbit.api.features.Holidays;
 import org.itsallcode.whiterabbit.api.features.MonthDataStorage;
@@ -28,9 +29,9 @@ class PluginManagerTest
     private PluginRegistry pluginRegistryMock;
 
     @Mock
-    private PluginWrapper plugin1;
+    private AppPluginImpl plugin1;
     @Mock
-    private PluginWrapper plugin2;
+    private AppPluginImpl plugin2;
 
     @Mock
     Holidays holidays1;
@@ -58,7 +59,7 @@ class PluginManagerTest
     void getProjectReportExporterPlugins_noPluginAvailable()
     {
         simulatePlugins();
-        assertThat(pluginManager.getProjectReportExporterPlugins()).isEmpty();
+        assertThat(pluginManager.findPluginsSupporting(ProjectReportExporter.class)).isEmpty();
     }
 
     @Test
@@ -66,7 +67,7 @@ class PluginManagerTest
     {
         simulatePlugins(plugin1);
         when(plugin1.supports(ProjectReportExporter.class)).thenReturn(false);
-        assertThat(pluginManager.getProjectReportExporterPlugins()).isEmpty();
+        assertThat(pluginManager.findPluginsSupporting(ProjectReportExporter.class)).isEmpty();
     }
 
     @Test
@@ -74,24 +75,35 @@ class PluginManagerTest
     {
         simulatePlugins(plugin1);
         when(plugin1.supports(ProjectReportExporter.class)).thenReturn(true);
-        assertThat(pluginManager.getProjectReportExporterPlugins()).containsExactly("plugin1");
+        assertThat(pluginManager.findPluginsSupporting(ProjectReportExporter.class)).containsExactly(plugin1);
     }
 
     @Test
     void getAllFeatures()
     {
-        when(plugin1.getFeature(Holidays.class)).thenReturn(holidays1);
-        when(plugin2.getFeature(Holidays.class)).thenReturn(holidays2);
+        when(plugin1.getFeature(Holidays.class)).thenReturn(Optional.of(holidays1));
+        when(plugin2.getFeature(Holidays.class)).thenReturn(Optional.of(holidays2));
         when(plugin1.supports(Holidays.class)).thenReturn(true);
         when(plugin2.supports(Holidays.class)).thenReturn(true);
         simulatePlugins(plugin1, plugin2);
         assertThat(pluginManager.getAllFeatures(Holidays.class)).containsExactly(holidays1, holidays2);
     }
 
-    private void simulatePlugins(PluginWrapper... plugins)
+    @Test
+    void pluginReturnsEmptyOptional()
+    {
+        when(plugin1.getFeature(Holidays.class)).thenReturn(Optional.of(holidays1));
+        when(plugin2.getFeature(Holidays.class)).thenReturn(Optional.empty());
+        when(plugin1.supports(Holidays.class)).thenReturn(true);
+        when(plugin2.supports(Holidays.class)).thenReturn(true);
+        simulatePlugins(plugin1, plugin2);
+        assertThat(pluginManager.getAllFeatures(Holidays.class)).containsExactly(holidays1);
+    }
+
+    private void simulatePlugins(AppPluginImpl... plugins)
     {
         lenient().when(pluginRegistryMock.getAllPlugins()).thenReturn(asList(plugins));
-        for (final PluginWrapper plugin : plugins)
+        for (final AppPluginImpl plugin : plugins)
         {
             lenient().when(pluginRegistryMock.getPlugin(plugin.getId())).thenReturn(plugin);
         }
@@ -101,9 +113,8 @@ class PluginManagerTest
     void getProjectReportExporter_noPluginAvailable_throwsException()
     {
         simulatePlugins();
-        assertThatThrownBy(() -> pluginManager.getProjectReportExporter("unknown"))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("Plugin 'unknown' not found");
+
+        assertThat(pluginManager.getAllFeatures(ProjectReportExporter.class)).isEmpty();
     }
 
     @Test
@@ -111,9 +122,8 @@ class PluginManagerTest
     {
         simulatePlugins(plugin1);
         when(plugin1.supports(ProjectReportExporter.class)).thenReturn(false);
-        assertThatThrownBy(() -> pluginManager.getProjectReportExporter("plugin1"))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("Plugin 'plugin1' does not support feature " + ProjectReportExporter.class.getName());
+
+        assertThat(pluginManager.getAllFeatures(ProjectReportExporter.class)).isEmpty();
     }
 
     @Test
@@ -122,9 +132,9 @@ class PluginManagerTest
         simulatePlugins(plugin1);
         when(plugin1.supports(ProjectReportExporter.class)).thenReturn(true);
         final ProjectReportExporter featureMock = mock(ProjectReportExporter.class);
-        when(plugin1.getFeature(ProjectReportExporter.class)).thenReturn(featureMock);
+        when(plugin1.getFeature(ProjectReportExporter.class)).thenReturn(Optional.of(featureMock));
 
-        assertThat(pluginManager.getProjectReportExporter("plugin1")).isNotNull().isSameAs(featureMock);
+        assertThat(pluginManager.getAllFeatures(ProjectReportExporter.class)).hasSize(1).contains(featureMock);
     }
 
     @Test
@@ -133,9 +143,9 @@ class PluginManagerTest
         simulatePlugins(plugin1);
         when(plugin1.supports(MonthDataStorage.class)).thenReturn(true);
         final MonthDataStorage featureMock = mock(MonthDataStorage.class);
-        when(plugin1.getFeature(MonthDataStorage.class)).thenReturn(featureMock);
+        when(plugin1.getFeature(MonthDataStorage.class)).thenReturn(Optional.of(featureMock));
 
-        assertThat(pluginManager.getMonthDataStorage()).isPresent().containsSame(featureMock);
+        assertThat(pluginManager.getUniqueFeature(MonthDataStorage.class)).isPresent().containsSame(featureMock);
     }
 
     @Test
@@ -144,7 +154,7 @@ class PluginManagerTest
         simulatePlugins(plugin1);
         when(plugin1.supports(MonthDataStorage.class)).thenReturn(false);
 
-        assertThat(pluginManager.getMonthDataStorage()).isEmpty();
+        assertThat(pluginManager.getUniqueFeature(MonthDataStorage.class)).isEmpty();
     }
 
     @Test
@@ -154,7 +164,7 @@ class PluginManagerTest
         when(plugin1.supports(MonthDataStorage.class)).thenReturn(true);
         when(plugin2.supports(MonthDataStorage.class)).thenReturn(true);
 
-        assertThatThrownBy(() -> pluginManager.getMonthDataStorage())
+        assertThatThrownBy(() -> pluginManager.getUniqueFeature(MonthDataStorage.class))
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessage(
                         "Found multiple plugins supporting " + MonthDataStorage.class.getName()
@@ -164,7 +174,7 @@ class PluginManagerTest
     @Test
     void getAllPlugins()
     {
-        final Collection<PluginWrapper> plugins = new ArrayList<>();
+        final Collection<AppPluginImpl> plugins = new ArrayList<>();
         when(pluginRegistryMock.getAllPlugins()).thenReturn(plugins);
         assertThat(pluginManager.getAllPlugins()).isSameAs(plugins);
     }